Fixing some implicit casts with data loss.
git-svn-id: https://svn.code.sf.net/p/previous/code/branches/branch_filesharing@1567 9ba66e4b-2233-44b8-a46a-7e93b2105c9e

+++ src/dimension/nd_sdl.cpp
@@ -116,7 +116,7 @@ void NDSDL::destroy(void) {
 
 void NDSDL::resize(float scale) {
     if (ndWindow) {
-        SDL_SetWindowSize(ndWindow, SDL_lroundf(scale*1120), SDL_lroundf(scale*832));
+        SDL_SetWindowSize(ndWindow, (int)SDL_lroundf(scale*1120), (int)SDL_lroundf(scale*832));
     }
 }
